+++ src/main/java/com/lowdragmc/lowdraglib2/gui/texture/DynamicTexture.java
@@ -21,7 +21,7 @@ public class DynamicTexture implements IGuiTexture {
 
     @Override
     @OnlyIn(Dist.CLIENT)
-    public void draw(GuiGraphics graphics, int mouseX, int mouseY, float x, float y, float width, float height, float partialTicks) {
+    public void draw(GuiGraphics graphics, float mouseX, float mouseY, float x, float y, float width, float height, float partialTicks) {
         textureSupplier.get().draw(graphics, mouseX, mouseY, x, y, width, height, partialTicks);
     }
 }
